+++ test/unit/fuzz-tools.test.js
@@ -343,6 +343,7 @@ import * as shafferAngleV363 from '../../lib/shaffer-angle-v363.js';
 import * as casTedV364 from '../../lib/cas-ted-v364.js';
 import * as pragueBarrettV365 from '../../lib/prague-barrett-v365.js';
 import * as neckZoneV366 from '../../lib/neck-zone-v366.js';
+import * as pasSwallowV367 from '../../lib/pas-swallow-v367.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -631,6 +632,7 @@ const MODULES = {
   'cas-ted-v364.js': casTedV364,
   'prague-barrett-v365.js': pragueBarrettV365,
   'neck-zone-v366.js': neckZoneV366,
+  'pas-swallow-v367.js': pasSwallowV367,
   'tb-testing.js': tbTesting,
 };